thread-safety: create a getter and remove GUARDED_BY
There is no obvious way for a data member to be make as guarded by a mutex for
writes and a shared_mutex for reads.  Removing GUARDED_BY is regrettable, but
marking the only writer functions as REQUIRES(mu_) should provide adequate
protection.  Make the shared_mutex mutable so that const functions can change it
and only it.