write a multithreaded binary to exercise the functions
Create a binary to call the functions and run it with TSAN.  Since the mutex
implementation is backed by C++'s std::mutex, TSAN recognizes that concurrent
access attempts are protected by them.   Why that is the case with this code and
not with librtpi's mutex implementation is not entirely clear.